+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/mfa/setup/SetupMfaCommand.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.command.user.mfa;
+package io.ciphernance.identity.application.command.user.mfa.setup;
 
 import io.ciphernance.identity.application.mediator.Command;
