+++ services/identity-service/src/main/java/io/ciphernance/identity/application/command/user/mfa/setup/SetupMfaResponse.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.command.user.mfa;
+package io.ciphernance.identity.application.command.user.mfa.setup;
 
 import java.util.UUID;
